updates BedrockIT, OpenAiAnthropicIT, PromptIT and VertexAiIT
